Make CS K8s debug jobs its own helm chart and deploy step
We make the CS K8s debug jobs its own Helm chart and deploy
step. We configure the new pipeline step of cs-debug-jobs
and we make it dependent on the CS deployment step. This
means the cs-debug-jobs pipeline step should only
be executed if the CS deployment step completes and
completes successfully.

Additionally, we've reorganized the Helm Charts directory
structure to conceptually separate them.

Finally, we also ensure that the template files within the
cs-debug-jobs Helm chart are rendered when the deployDebugJob
helm parameter is set to true. This allows us to conditionally
deploy the elements.We configure the new pipeline step of cs-debug-jobs
and we make it dependent on the CS deployment step. This
means the cs-debug-jobs pipeline step should only
be executed if the CS deployment step completes and
completes successfully.